moved ModalCreatePlaylist Component into jsx file and called it in Room.jsx, activated by buttons in Myplaylists and MediaEntry. Note, there is a checkbox bug where the checkbox doesn't toggle if default is checked...
