+++ public/javascripts/script.js
@@ -220,6 +220,19 @@ $('#create-room-toggle').change(function(){
     };
 });
 
+$('#create-playlist-toggle').change(function(){
+    if (this.checked) {
+        console.log('this is checked');
+        $('#create-playlist-toggle-icon').removeClass('fa-globe');
+        $('#create-playlist-toggle-icon').addClass('fa-lock');
+    }
+    else {
+        console.log('this is NOT checked');
+        $('#create-playlist-toggle-icon').removeClass('fa-lock');
+        $('#create-playlist-toggle-icon').addClass('fa-globe');
+    };
+});
+
 
 /* Mobile Tab Navigation */
 
@@ -284,4 +297,3 @@ $('#mobile-tab-search').click(function() {
     $('.tabbed-container-mobile-collapse').css('display','block');
     document.getElementById('tab-search').click();
 });
-
